renamed CarInfoClass
